+++ HISTORY.md
@@ -6,6 +6,7 @@
     * export script (including dependencies) via its context menu
     * export / import sprite-local custom block definitions from the palette
     * added "combinations" primitive to the palette
+    * new POSITION primitive reporter in the MOTION category
     * new "position" choice in OF reporter's attribute dropdown, reports a list of XY coordinates
     * new "Tad" costume series, thanks, Meghan and Brian!
 * **Notable Changes:**
@@ -31,6 +32,7 @@
 * threads, objects: new menu functionality for ASK command, when passing a list
 * objects: support various data types inside menus (sprites, costumes, blocks)
 * objects: wrap long texts inside menus into several lines
+* objects: added new POSITION primitive block to the MOTION category
 
 ### 2022-03-28
 * new "Tad" turtle costumes, thanks, Meghan and Brian!